requirements.txt
Django>=5.0
Telethon
uvicorn[standard]
python-dotenv
cryptg